DevOps: fixes for aiida-qe v4.5.x
The aiida-quantumespresso v4.5.x comes with the `qpoints_distance`
input for the `PhBaseWorkChain`. This broke the compatibility with
the current code, as the `ph.qpoints` input has been promoted to
an higher level in the namespace, i.e. to just `qpoints`.
Some smal adjustments related to the tests and concerning the
new aiida-qe version are also applied.
The aiida-qe dependency is upgraded to be >=4.5.